requirements.txt
pycryptodome
huffman